--- src/main/java/ch/zuegi/ordermgmt/feature/test/AuthorId.java
@@ -1,10 +1,0 @@
-package ch.zuegi.ordermgmt.feature.test;
-
-import ch.zuegi.ordermgmt.shared.RandomUUID;
-
-public class AuthorId  extends RandomUUID {
-    @Override
-    protected String getPrefix() {
-        return "AUT-%s";
-    }
-}
