--- src/main/java/ch/zuegi/ordermgmt/feature/test/CreateArticle.java
@@ -1,11 +1,0 @@
-package ch.zuegi.ordermgmt.feature.test;
-
-import ch.zuegi.ordermgmt.shared.Command;
-import lombok.Value;
-
-
-@Value(staticConstructor = "commandOf")
-public class CreateArticle implements Command {
-    String title;
-    CreateAuthor author;
-}
