--- src/main/java/ch/zuegi/ordermgmt/feature/test/shared/DomainCommandHandler.java
@@ -1,8 +1,0 @@
-package ch.zuegi.ordermgmt.feature.test.shared;
-
-import ch.zuegi.ordermgmt.shared.Command;
-
-public interface DomainCommandHandler<A, C extends Command>  {
-     void handle(C command);
-
-}
