+++ src/main/java/ch/zuegi/ordermgmt/feature/ticket/domain/TicketPositionAdded.java
@@ -1,9 +1,9 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain;
+package ch.zuegi.ordermgmt.feature.ticket.domain;
 
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.shared.DomainEvent;
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo.TicketId;
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo.TicketPositionId;
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo.TradeItemId;
+import ch.zuegi.ordermgmt.feature.ticket.domain.vo.TicketId;
+import ch.zuegi.ordermgmt.feature.ticket.domain.vo.TicketPositionId;
+import ch.zuegi.ordermgmt.feature.ticket.domain.vo.TradeItemId;
+import ch.zuegi.ordermgmt.shared.DomainEvent;
 import lombok.Value;
 
 import java.math.BigDecimal;
