+++ src/main/java/ch/zuegi/ordermgmt/feature/ticket/domain/TicketPositionIdStartWithException.java
@@ -1,4 +1,4 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain;
+package ch.zuegi.ordermgmt.feature.ticket.domain;
 
 public class TicketPositionIdStartWithException extends RuntimeException{
     public TicketPositionIdStartWithException(String s) {
