--- src/main/java/ch/zuegi/ordermgmt/feature/ticket/domain/command/CreateOrder.java
@@ -1,16 +1,0 @@
-package ch.zuegi.ordermgmt.feature.ticket.domain.command;
-
-import ch.zuegi.ordermgmt.feature.ticket.domain.entity.TradeDealType;
-import ch.zuegi.ordermgmt.shared.Command;
-import lombok.Value;
-
-import java.util.List;
-
-@Value(staticConstructor = "commandOf")
-public class CreateOrder implements Command {
-
-    TradeDealType tradeDealType;
-
-    List<CreateOrderPosition> createOrderPositionList;
-
-}
