--- src/main/java/ch/zuegi/ordermgmt/feature/ticket/domain/entity/OrderEntity.java
@@ -1,17 +1,0 @@
-package ch.zuegi.ordermgmt.feature.ticket.domain.entity;
-
-import ch.zuegi.ordermgmt.feature.ticket.domain.vo.OrderId;
-import ch.zuegi.ordermgmt.shared.Entity;
-
-public class OrderEntity extends Entity<OrderId> {
-
-
-    public OrderEntity(OrderId aggregateId) {
-        super(aggregateId);
-    }
-
-    @Override
-    public OrderId id() {
-        return aggregateId;
-    }
-}
