--- src/main/java/ch/zuegi/ordermgmt/feature/vaughn/ticket/domain/shared/ValueObject.java
@@ -1,15 +1,0 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.shared;
-
-import java.io.Serializable;
-
-public interface ValueObject<T> extends Serializable {
-
-    /**
-     * Value objects compare by the values of their attributes, they don't have an identity.
-     *
-     * @param other The other value object.
-     * @return <code>true</code> if the given value object's and this value object's attributes are the same.
-     */
-    boolean sameValueAs(T other);
-
-}
