--- src/main/java/ch/zuegi/ordermgmt/feature/vaughn/ticket/domain/vo/TradeItemId.java
@@ -1,20 +1,0 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo;
-
-import ch.zuegi.ordermgmt.shared.EventPrefix;
-import ch.zuegi.ordermgmt.shared.RandomUUID;
-
-public class TradeItemId extends RandomUUID {
-
-    public TradeItemId() {
-        super();
-    }
-
-    public TradeItemId(String id) {
-        super(id);
-    }
-
-    @Override
-    protected String getPrefix() {
-        return EventPrefix.TIP.name()+ "-%s";
-    }
-}
