+++ src/main/java/ch/zuegi/ordermgmt/shared/DomainEvent.java
@@ -1,4 +1,4 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.shared;
+package ch.zuegi.ordermgmt.shared;
 
 import java.time.LocalDateTime;
