--- src/main/java/ch/zuegi/ordermgmt/shared/aggregateRoot/AggregateRootBehaviorBuilder.java
@@ -1,23 +1,0 @@
-package ch.zuegi.ordermgmt.shared.aggregateRoot;
-
-import ch.zuegi.ordermgmt.shared.Command;
-import ch.zuegi.ordermgmt.shared.CommandHandler;
-import ch.zuegi.ordermgmt.shared.Entity;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class AggregateRootBehaviorBuilder<RandomUUID> {
-
-    private final Map<Class<? extends Command>, CommandHandler<? extends Command, ? extends Entity<RandomUUID>, RandomUUID>> commandHandlers = new HashMap<>();
-
-    public AggregateRootBehavior<RandomUUID> build() {
-        return new AggregateRootBehavior<>(commandHandlers);
-    }
-
-
-    public <A extends Command, B extends Entity<RandomUUID>> void setCommandHandler(Class<? extends Command> commandClass, CommandHandler<? extends Command, ? extends Entity<RandomUUID>, RandomUUID> handler) {
-        commandHandlers.put( commandClass,  handler);
-    }
-
-}
