+++ src/test/java/ch/zuegi/ordermgmt/feature/ticket/domain/DomainTest.java
@@ -1,4 +1,4 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain;
+package ch.zuegi.ordermgmt.feature.ticket.domain;
 
 public abstract class DomainTest extends AbstractEventTrackingTest {
