+++ src/test/java/ch/zuegi/ordermgmt/feature/ticket/domain/TicketTest.java
@@ -1,10 +1,9 @@
-package ch.zuegi.ordermgmt.feature.vaughn.ticket.domain;
+package ch.zuegi.ordermgmt.feature.ticket.domain;
 
 
-import ch.zuegi.ordermgmt.feature.ticket.domain.Order;
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo.TicketId;
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo.TicketPositionId;
-import ch.zuegi.ordermgmt.feature.vaughn.ticket.domain.vo.TradeItemId;
+import ch.zuegi.ordermgmt.feature.ticket.domain.vo.TicketId;
+import ch.zuegi.ordermgmt.feature.ticket.domain.vo.TicketPositionId;
+import ch.zuegi.ordermgmt.feature.ticket.domain.vo.TradeItemId;
 import ch.zuegi.ordermgmt.shared.aggregateRoot.AggregateRootValidationException;
 import ch.zuegi.ordermgmt.shared.aggregateRoot.AggregateRootValidationMsg;
 import org.assertj.core.api.Assertions;
